feat(youtube): add advanced player controls
